Add files via upload
In Java, a String is an object that represents a sequence of characters. Java provides a robust and flexible API for handling strings, allowing for various operations such as concatenation, comparison, and manipulation. In this article, we will go through the Java String concept in detail.

What are Strings in Java?
Strings are the type of objects that can store the character of values and in Java, every character is stored in 16 bits i,e using UTF 16-bit encoding. A string acts the same as an array of characters in Java.
Ways of Creating a String
There are two ways to create a string in Java: 

String Literal
Using new Keyword
String literal
To make Java more memory efficient (because no new objects are created if it exists already in the string constant pool). 

Example:
string  literal
String s= “kaushalshinde910”;
Using new keyword
String s = new String (“kaushalshinde910”);